refactor: move the PyTensor translator into a module without elicito imports
